n-acd: add a delay between announcements
NM passes a zero timeout to n-acd to signal that probe should be
skipped. This has the side effect of setting the timeout-multiplier to
zero and thus every timeout becomes zero as well. Instead, set a fixed
non-null value when announcing. 4500 means that we use half of the
intervals stated in RFC 5227, that is 1 second between announcement
(the same value used by arping).