all: add stable-id specifiers "${HOST}" and "${DEVICE}"
Add two new stable-id specifiers "${HOST}" and "${DEVICE}" to explicitly
declare that the connection's identity differs per-host or per-device.

Note that for settings like "ipv6.addr-gen-mode=stable" we already hash
the host's secret-key and the interface's name. So, for addr-gen-mode,
using these specifiers has no real use. But for example, we don't do that for
"ipv4.dhcp-client-id=stable". Currently, "ipv4.dhcp-client-id=stable" only
hashes the stable-id alone. The latter is a bug, we really want to generate
per-host client-ids and will be fixed next.

The point is, in various context we possibly anyway already include a
per-host or a per-device token into the generation algorithm. However,
the new specifiers are still useful, to declare this concept of a
per-host or per-device token in the stable-id itself.

While arguably "${HOST}" is (soon) anyway implicitly used everywhere,
"${DEVICE}" is genuinely useful. Especially for the DHCPv4 client identifier
which is supposed to differ between interfaces (according to RFC). While
we don't do that by default with "ipv4.dhcp-client-id=stable", it is
now possible. Note that the fact that the client-id by default is the
same accross interfaces, is also not a common problem, because commonly
profiles are restricted to one device via connection.interface-name.
Now the user can create profiles that activate on multiple
devices, and configurably use the same or different stable client-id.